+++ cmd/whisprgo/main.go
@@ -4,18 +4,18 @@ import (
 	"fmt"
 	"os"
 
-	"whispergo/internal/app"
+	"whisprgo/internal/app"
 )
 
 func main() {
 	application, err := app.New()
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "whispergo: %v\n", err)
+		fmt.Fprintf(os.Stderr, "whisprgo: %v\n", err)
 		os.Exit(1)
 	}
 
 	if err := application.Execute(); err != nil {
-		fmt.Fprintf(os.Stderr, "whispergo: %v\n", err)
+		fmt.Fprintf(os.Stderr, "whisprgo: %v\n", err)
 		os.Exit(1)
 	}
 }